# Use static property names type