--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -10,17 +10,20 @@ jobs:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }}
     runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: true
+      fail-fast: false
       matrix:
         version:
           - '1.0'
+          - '1.1'
+          - '1.2'
           - '1.3'
+          - '1.4'
           - '1.5'
+          - '1.6'
+          - '1.7'
           - 'nightly'
         os:
           - ubuntu-latest
-          - macOS-latest
-          - windows-latest
         arch:
           - x64
     steps: