--- a/src/ConstructionBase.jl
+++ b/src/ConstructionBase.jl
@@ -37,10 +37,6 @@ struct PropertyIndices{inds} end
 
 const PropertyKeys{K} = Union{PropertyNames{K},PropertyIndices{K}}
 
-Base.eltype(@nospecialize p::PropertyKeys) = eltype(typeof(p))
-Base.eltype(@nospecialize T::Type{<:PropertyIndices}) = Int
-Base.eltype(@nospecialize T::Type{<:PropertyNames}) = Symbol
-
 PropertyNames(::Type{T}) where {T} = PropertyNames{fieldnames(T)}()
 PropertyNames(@nospecialize(T::Type{<:NamedTuple})) = PropertyNames{T.parameters[1]}()
 PropertyNames(@nospecialize(x)) = PropertyNames(typeof(x))
@@ -67,12 +63,8 @@ end
 function Base.getindex(@nospecialize(p1::PropertyNames), @nospecialize(p2::PropertyIndices))
     PropertyNames{_getfields(values(p1), values(p2), _vlength(p2))}()
 end
-
-Base.isdone(@nospecialize(p::PropertyKeys), i::Int) = length(p) < i
-Base.iterate(::PropertyKeys{()}) = nothing
-Base.iterate(@nospecialize(p::PropertyKeys)) = @inbounds(p[1]), 2
-@inline function Base.iterate(@nospecialize(p::PropertyKeys), state::Int)
-    if Base.isdone(p, state)
+@inline function Base.iterate(@nospecialize(p::PropertyKeys), state::Int=1)
+    if length(p) > state
         return nothing
     else
         return @inbounds(p[state]), state + 1
@@ -114,15 +106,15 @@ end
 ##### setproperties
 ################################################################################
 setproperties(obj; kw...) = setproperties(obj, (;kw...))
-setproperties(obj             , patch::Tuple      ) = setproperties_object(obj     , patch )
-setproperties(obj             , patch::NamedTuple ) = setproperties_object(obj     , patch )
-setproperties(obj::NamedTuple , patch::Tuple      ) = setproperties_namedtuple(obj , patch )
-setproperties(obj::NamedTuple , patch::NamedTuple ) = setproperties_namedtuple(obj , patch )
-setproperties(obj::Tuple      , patch::Tuple      ) = setproperties_tuple(obj      , patch )
-setproperties(obj::Tuple      , patch::NamedTuple ) = setproperties_tuple(obj      , patch )
-
-setproperties_namedtuple(obj, patch::Tuple{}) = obj
-@noinline function setproperties_namedtuple(obj, patch::Tuple)
+setproperties(obj, patch) = isempty(patch) ? obj : _setproperties(obj, patch)
+function _setproperties(@nospecialize(obj::Tuple), @nospecialize(patch::Tuple))
+    _generated_setproperties(obj, patch, PropertyIndices(obj), PropertyIndices(patch))
+end
+function _setproperties(obj, patch)
+    _generated_setproperties(obj, patch, PropertyNames(obj), PropertyNames(patch))
+end
+_setproperties(::NamedTuple{fields}, patch::NamedTuple{fields}) where {fields} = patch
+@noinline function _setproperties(@nospecialize(obj::NamedTuple), patch::Tuple)
     msg = """
     setproperties(obj::NamedTuple, patch::Tuple) only allowed for empty Tuple. Got:
     obj = $obj
@@ -130,33 +122,14 @@ setproperties_namedtuple(obj, patch::Tuple{}) = obj
     """
     throw(ArgumentError(msg))
 end
-function setproperties_namedtuple(@nospecialize(obj), @nospecialize( patch))
-    _generated_setproperties(obj, patch, PropertyNames(obj), PropertyNames(patch))
-end
-function validate_setproperties_result(
-    nt_new::NamedTuple{fields}, nt_old::NamedTuple{fields}, obj, patch) where {fields}
-    nothing
-end
-@noinline function validate_setproperties_result(nt_new, nt_old, obj, patch)
-    O = typeof(obj)
-    _setproperties_object(O, fieldnames(O), fieldnames(typeof(patch)))
-end
-@noinline function _setproperties_result_error(O::DataType, @nospecialize(objsyms), @nospecialize(patchsyms))
+@noinline function _setproperties(@nospecialize(obj), @nospecialize(patch::Tuple))
     msg = """
-    Failed to assign properties $(patchsyms) to object with fields $(objsyms).
-    You may want to overload
-    ConstructionBase.setproperties(obj::$O, patch::NamedTuple)
-    ConstructionBase.getproperties(obj::$O)
+    setproperties(obj, patch::Tuple) only allowed for empty Tuple. Got:
+    obj = $obj
+    patch = $patch
     """
-    throw(ArgumentError(msg))
 end
-
-function setproperties_namedtuple(::NamedTuple{fields}, patch::NamedTuple{fields}) where {fields}
-    patch
-end
-
-setproperties_tuple(obj::Tuple, patch::NamedTuple{()}) = obj
-@noinline function setproperties_tuple(obj::Tuple, patch::NamedTuple)
+@noinline function _setproperties(@nospecialize(obj::Tuple), @nospecialize(patch::NamedTuple))
     msg = """
     setproperties(obj::Tuple, patch::NamedTuple) only allowed for empty NamedTuple. Got:
     obj  ::Tuple      = $obj
@@ -164,57 +137,31 @@ setproperties_tuple(obj::Tuple, patch::NamedTuple{()}) = obj
     """
     throw(ArgumentError(msg))
 end
-setproperties_tuple(::NTuple{N,Any}, patch::NTuple{N,Any}) where {N} = patch
-append(x,y) = (x..., y...)
-@noinline function throw_setproperties_tuple_error(obj, patch)
-    msg = """
-    Cannot call `setproperties(obj::Tuple, patch::Tuple)` with `length(obj) < length(patch)`. Got:
-    obj = $obj
-    patch = $patch
-    """
-    throw(ArgumentError(msg))
-end
-function setproperties_tuple(obj::NTuple{N,Any}, patch::NTuple{K,Any}) where {N,K}
-    if K > N
-        throw_setproperties_tuple_error(obj, patch)
-    end
-    append(patch, after(obj, Val{K}()))
-end
-function after(xs::Tuple, ::Val{N}) where {N}
-    after(Base.tail(xs), Val{N-1}())
-end
-function after(x::Tuple, ::Val{0})
-    x
-end
-
-setproperties_object(obj, patch::Tuple{}) = obj
-@noinline function setproperties_object(obj, @nospecialize(patch::Tuple))
-    msg = """
-    setproperties(obj, patch::Tuple) only allowed for empty Tuple. Got:
-    obj = $obj
-    patch = $patch
-    """
-end
-setproperties_object(obj, patch::NamedTuple{()}) = obj
-@inline function setproperties_object(obj, patch)
-    _generated_setproperties(obj, patch, PropertyNames(obj), PropertyNames(patch))
-end
-@generated function _generated_setproperties(obj, patch, ::PropertyNames{objsyms}, ::PropertyNames{patchsyms}) where {objsyms,patchsyms}
+@generated function _generated_setproperties(obj, patch, ::PropertyKeys{objkeys}, ::PropertyKeys{patchkeys}) where {objkeys,patchkeys}
     out = Expr(:call, :(constructorof(typeof(obj))))
-    names = Symbol[objsyms...]
-    for ps in patchsyms
-        if !in(ps, objsyms)
+    names = [objkeys...]
+    for ps in patchkeys
+        if !in(ps, objkeys)
             push!(names, ps)
         end
     end
-    if all(in(objsyms), names)
+    if all(in(objkeys), names)
         for n in names
             qn = QuoteNode(n)
-            push!(out.args, in(n, patchsyms) ? Expr(:call, :getfield, :patch, qn) : Expr(:call, :getproperty, :obj, qn))
+            push!(out.args, in(n, patchkeys) ? Expr(:call, :getfield, :patch, qn) : Expr(:call, :getproperty, :obj, qn))
         end
         return out
     else
-        return :(_setproperties_result_error(typeof(obj), objsyms, patchsyms))
+        return quote
+            O = typeof(obj)
+            msg = """
+            Failed to assign properties $(patchkeys) to object with fields $(objkeys).
+            You may want to overload
+            ConstructionBase.setproperties(obj::$O, patch::NamedTuple)
+            ConstructionBase.getproperties(obj::$O)
+            """
+            throw(ArgumentError(msg))
+        end
     end
 end