--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -35,7 +35,16 @@ end
     x = (a = 1, b  = 2, c = 3, d = 4)
     @test @inferred(getproperties(x, abprops)) === (a = 1, b = 2)
 end
-
+@testset "Issue #55" begin
+    struct Issue55
+        x::Int
+        y::Union{Nothing, Int}
+    end 
+    function update(x)
+        setproperties(x, (; x = x.x + 1))
+    end
+    @inferred update(Issue55(1,2))
+end
 @testset "getproperties" begin
     o = AB(1, 2)
     @test getproperties(o) === (a=1, b=2)