I was trying to put something together for Accessors.jl and found that some changes here can really improve performance downstream. 

The strategy used here is to only depend on `PropertyNames{fieldnames(obj)}()` to explicitly inject compile time information.
Only using `PropertyNames{names}` to construct the body of generated methods is more reliable than generic `::Type{T}`.

Using the example from [this isssue](https://github.com/JuliaObjects/ConstructionBase.jl/issues/55) demonstrates some of the performance differences here.

Where `A` and `B` are immutable types and `B` has a field with a union, this is what I get on the current master branch.
```julia
julia> @btime update($(Ref(A(1, 2)))[]);
  1.471 ns (0 allocations: 0 bytes)

julia> @btime update($(Ref(B(1, 2)))[]);
  387.926 ns (6 allocations: 176 bytes)
```

Problems with the union field are mostly absent using this PR.
```julia
julia> @btime update($(Ref(A(1, 2)))[]);
  1.760 ns (0 allocations: 0 bytes)

julia> @btime update($(Ref(B(1, 2)))[]);
  8.819 ns (0 allocations: 0 bytes)
```

Explicitly defining `PropertyNames` helps gives a tiny boost to performance.
```julia
ConstructionBase.PropertyNames(::Type{A}) = ConstructionBase.PropertyNames{(:x, :y)}()
ConstructionBase.PropertyNames(::Type{B}) = ConstructionBase.PropertyNames{(:x, :y)}()

julia> @btime update($(Ref(A(1, 2)))[]);
  1.471 ns (0 allocations: 0 bytes)

julia> @btime update($(Ref(B(1, 2)))[]);
  8.034 ns (0 allocations: 0 bytes)
``.`

This approach also allows manually unrolling operations so bigger collections get big performance boosts.
```julia
x = ntuple(identity, 40);
y = ntuple(identity, 30);

# master
@btime setproperties($x, $y); # 10.250 μs (10 allocations: 3.00 KiB)

# this PR
@btime setproperties($x, $y); # 3.984 ns (0 allocations: 0 bytes)
```

It might be worth noting that another internal change to `setproperties` was removing use of `getproperties`.
`PropertyNames` provides the information we really need from `getproperties` for reconstruction, which in turn helps us avoid needless `getproperty` calls.

I haven't changed any tests or documented anything here yet because I thought there might be some input on how well this really fits this package or if significant amount needs to be changed.