@Tokazama 2022-04-29
Sounds good. I was a bit worried because I ended up changing more than I originally planned to the source code, but none of the tests have been changed so it shouldn't break anything. Theoretically, we could say that the only methods related to this package that should ever be overloaded are `constructorof`, `PropertyNames`, and `getproperty`. I don't think there's a situation where users would get better performance by defining a unique `getproperties` or `setproperties` method given these other definitions, but I didn't want to overstep my bounds here.

I haven't taken the time to ensure that start up times are good but the use of `@generated` here is so minimal I don't think it will be an issue.