@jw3126 2022-05-02
> Theoretically, we could say that the only methods related to this package that should ever be overloaded are constructorof, PropertyNames, and getproperty

One usecase that comes up for me from time to time is like this. I have a file containing a huge vector of objects as "packed bits" and I want to mmap it. For this I do something like:
```julia
struct MyObject
    data::NTuple{17, UInt8}
end
function ConstructionBase.getproperties(o::MyObject)
    ...
end
function ConstructionBase.setproperties(o::MyObject, patch::NamedTuple)
    ...
end
```