@jw3126 2022-05-02
In my use cases, the property names are statically known and hard coded in the `getproperties` / `setproperties` functions.
Here is a [simple](https://github.com/jw3126/PhaseSpaceIO.jl/blob/master/src/egs/particle.jl#L75) and [complex](
https://github.com/jw3126/PhaseSpaceIO.jl/blob/master/src/iaea/compressed.jl#L175) example.