@Tokazama 2022-05-02
Is the problem there that you aren't constructing `Latch` with positional arguments?