@jw3126 2022-05-02
Yes, I would prefer not to add a positional constructor like `Latch(multicross,charge,creation,visited,brems)`. Generally we would like to use `getproperties/setproperties` for fancy custom properties and `constructorof` for construction from `raw` private struct data. See also this discussion: https://github.com/JuliaObjects/ConstructionBase.jl/issues/53