@Tokazama 2022-05-02
With the current code it works like this

```julia
function setproperties_object(obj, patch)
    nt = getproperties(obj)
    nt_new = merge(nt, patch)
    validate_setproperties_result(nt_new, nt, obj, patch)
    constructorof(typeof(obj))(Tuple(nt_new)...)
end

```

Is the assumption that if one defines a custom `getproperties` then they also define a custom `setproperties`? It doesn't really change the core aspect of what's here but I want to make sure I understand so I can properly document things around what I've done.

It's a bit orthogonal to this PR, but would it make sense to have the following:

* Property based
  * `construct_from_properties`
  * `getproperties`
  * `setproperties`
* Field based
  * `constructorof`
  * `getfields`
  * `setfields`

And the field based ones would be completely determined by the type system