@Tokazama 2022-05-02
Sounds good to me. Would you like me to put a doc strings in here. I'm not sure how you want it structured