@jw3126 2022-05-02
Docstrings for private functions are not mandatory, but feel free to provide some. I guess supporting julia 1.0 will be painful, but it would be nice if we can support 1.6. Also, some inference tests that pass with this PR but not without it would be nice.