@jw3126 2022-05-02
We have some really long docstrings in `.md` files, but short docstrings can be placed in `.jl` files IMO.