@Tokazama 2022-05-02
Should `PropertyNames` be part of the public API because it can be passed to `getproperties` or or do we not want to actually support that feature?