@jw3126 2022-05-04
I think we should not make `getproperties(obj, ::PropertyNames)` public. The reason is that it makes overloading of `getproperties` harder for users.