@jw3126 2022-05-04
CI segfaults https://github.com/JuliaObjects/ConstructionBase.jl/runs/6290163099?check_suite_focus=true#step:5:68

Locally tests pass for me on 1.7