@Tokazama 2022-05-04
Passes locally on 1.7 and 1.8 for me.