@jw3126 2022-05-06
Maybe the segfault is related to `@pure`? Maybe you can restore the old `@generated` definition of `constructorof` in case `@assume_effects` is not available?