@Tokazama 2022-10-23
I kind of forgot about it. I think there are some inference failures on 1.6 still