@jw3126 2023-07-17
I think the inference issues addressed by this PR no longer exist. Feel free to reopen otherwise.